requirements.txt
fastapi==0.115.0
uvicorn==0.30.0

nltk==3.9.1
scikit-learn==1.5.1
numpy==1.26.4

sentence-transformers==2.7.0
torch==2.3.1

pdfplumber==0.11.4
python-docx==1.1.2

python-multipart==0.0.9